Module Sexp_pretty.Config

type color =

| Black

| Red

| Green

| Yellow

| Blue

| Magenta

| Cyan

| White

| Default

val sexp_of_color : color -> Sexplib0.Sexp.t

val color_of_sexp : Sexplib0.Sexp.t -> color

type atom_threshold =

| Atom_threshold of Base.int

Datatypes of various thresholds

val sexp_of_atom_threshold : atom_threshold -> Sexplib0.Sexp.t

val atom_threshold_of_sexp : Sexplib0.Sexp.t -> atom_threshold

type char_threshold =

| Character_threshold of Base.int

val sexp_of_char_threshold : char_threshold -> Sexplib0.Sexp.t

val char_threshold_of_sexp : Sexplib0.Sexp.t -> char_threshold

type depth_threshold =

| Depth_threshold of Base.int

Depth is the depth of an atom. For example, in (a (b (c) d)), the depth of a is 1, the depth of b and d is 2, and depth of c is 3. Depth_threshold usually refers to the maximum depth of any atom in a list for it to be considered for certain heuristic, e.g. data alignment.

val sexp_of_depth_threshold : depth_threshold -> Sexplib0.Sexp.t

val depth_threshold_of_sexp : Sexplib0.Sexp.t -> depth_threshold

type aligned_parens =

| Parens_alignment of Base.bool

Whether or not should closing parentheses be aligned.

val sexp_of_aligned_parens : aligned_parens -> Sexplib0.Sexp.t

val aligned_parens_of_sexp : Sexplib0.Sexp.t -> aligned_parens

type data_alignment =

| Data_not_aligned

| Data_aligned of aligned_parens * atom_threshold * char_threshold * depth_threshold	(*
Character threshold is excluding spaces and parentheses, the maximum depth can't exceed the depth threshold.
*)

val sexp_of_data_alignment : data_alignment -> Sexplib0.Sexp.t

val data_alignment_of_sexp : Sexplib0.Sexp.t -> data_alignment

type atom_coloring =

| Color_first of Base.int	(*
Color the first one, only if the number of atoms that follow it at most the value of the constructor's argument.
*)

| Color_all

| Color_none

val sexp_of_atom_coloring : atom_coloring -> Sexplib0.Sexp.t

val atom_coloring_of_sexp : Sexplib0.Sexp.t -> atom_coloring

type comment_indent =

| Auto_indent_comment

| Indent_comment of Base.int

This currently relates only to block comments. Auto_indent tries to infer the indentation from the original formatting, Indent_comment n indents new lines in a block comment by n spaces.

val sexp_of_comment_indent : comment_indent -> Sexplib0.Sexp.t

val comment_indent_of_sexp : Sexplib0.Sexp.t -> comment_indent

type comment_print_style =

| Pretty_print	(*
Auto aligns multi-line block comments.
*)

| Conservative_print	(*
Leaves block comments as they are, only adjusts indentation.
*)

val all_of_comment_print_style : comment_print_style Base.list

val sexp_of_comment_print_style : comment_print_style -> Sexplib0.Sexp.t

val comment_print_style_of_sexp : Sexplib0.Sexp.t -> comment_print_style

type comments =

| Drop

| Print of comment_indent * color Base.option * comment_print_style

Comment treatment.

val sexp_of_comments : comments -> Sexplib0.Sexp.t

val comments_of_sexp : Sexplib0.Sexp.t -> comments

type atom_printing =

| Escaped	(*
Can be parsed again. Atoms are printed out as loaded, with escaping.
*)

| Minimal_escaping	(*
As Escaped, but applies escaping to fewer characters.
*)

| Interpreted	(*
Try to interpret atoms as sexps.
*)

val sexp_of_atom_printing : atom_printing -> Sexplib0.Sexp.t

val atom_printing_of_sexp : Sexplib0.Sexp.t -> atom_printing

type singleton_limit =

| Singleton_limit of atom_threshold * char_threshold

Singleton_lists are lists of the following format

(ATOM_1 .. ATOM_N (....))

and are printed in the following way if they are too big to fit on a line/force a breakline for other reasons:

(ATOM_1 .. ATOM_N ( .... ))

Thresholds correspond to what's an acceptable number/size of the leading atoms ATOM_1 through ATOM_N.

Character threshold is excluding spaces.

val sexp_of_singleton_limit : singleton_limit -> Sexplib0.Sexp.t

val singleton_limit_of_sexp : Sexplib0.Sexp.t -> singleton_limit

type paren_coloring = Base.bool

Should parentheses be colored?

val sexp_of_paren_coloring : paren_coloring -> Sexplib0.Sexp.t

val paren_coloring_of_sexp : Sexplib0.Sexp.t -> paren_coloring

type separator =

| No_separator

| Empty_line

Separator between individual sexps.

val sexp_of_separator : separator -> Sexplib0.Sexp.t

val separator_of_sexp : Sexplib0.Sexp.t -> separator

type parens =

| Same_line

| New_line

Should closing parentheses be on the same line as the last sexp in the list (modulo comments), or should they be on new lines? Should opening parentheses always be on the same line as what follows them, or should they sometimes (when the first item in the list is a list followed by some other sexp) be on a separate line?

val sexp_of_parens : parens -> Sexplib0.Sexp.t

val parens_of_sexp : Sexplib0.Sexp.t -> parens

type sticky_comments =

| Before

| Same_line

| After

Where to put line comments relative to an associated sexp.

val sexp_of_sticky_comments : sticky_comments -> Sexplib0.Sexp.t

val sticky_comments_of_sexp : Sexplib0.Sexp.t -> sticky_comments

type t = {

indent : Base.int;

data_alignment : data_alignment;

color_scheme : color Base.array;

atom_coloring : atom_coloring;

atom_printing : atom_printing;

paren_coloring : paren_coloring;

opening_parens : parens;

closing_parens : parens;

comments : comments;

singleton_limit : singleton_limit;

leading_threshold : atom_threshold * char_threshold;

separator : separator;

sticky_comments : sticky_comments;

}

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val default : t

val create : ?color:Base.bool -> ?interpret_atom_as_sexp:Base.bool -> ?drop_comments:Base.bool -> ?new_line_separator:Base.bool -> ?custom_data_alignment:data_alignment -> Base.unit -> t

val update : ?color:Base.bool -> ?interpret_atom_as_sexp:Base.bool -> ?drop_comments:Base.bool -> ?new_line_separator:Base.bool -> ?custom_data_alignment:data_alignment -> t -> t