Module Sexp_pretty.Sexp_with_layout

type 'a writer = Config.t -> 'a -> Sexplib.Sexp.With_layout.t_or_comment -> Base.unit

val pp_formatter : Caml.Format.formatter writer

pp_formatter conf fmt sexp will mutate the fmt with functions such as set_formatter_tag_functions

val pp_formatter' : next:( Base.unit -> Sexplib.Sexp.With_layout.t_or_comment Base.option ) -> Config.t -> Caml.Format.formatter -> Base.unit

val pp_buffer : Base.Buffer.t writer

val pp_out_channel : Caml.out_channel writer

val pp_blit : ( Base.string, Base.unit ) Base.Blit.sub writer

val pretty_string : Config.t -> Sexplib.Sexp.With_layout.t_or_comment -> Base.string

pretty_string needs to allocate. If you care about performance, using one of the pp_* functions above is advised.

val sexp_to_string : Sexplib.Sexp.With_layout.t_or_comment -> Base.string